final version
Here is the English version of the description:

The Piggy Bank project allows users to virtually manage their savings by adding money, shaking, breaking, and repairing a piggy bank. The application supports different types of banknotes and coins, simulating the process of adding each currency to the piggy bank. Additionally, it provides users with information about the fullness of the piggy bank.